Move OMEinsum code to package extension
