+++ Project.toml
@@ -8,7 +8,6 @@ Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 ArgCheck = "dce04be8-c92d-5529-be00-80e4d2c0e197"
 EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-OMEinsum = "ebe7aa44-baf0-506c-a96f-8464559b3922"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [weakdeps]
@@ -16,6 +15,7 @@ CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 Dagger = "d58978e5-989f-55fb-8d15-ea34adc7bf54"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
+OMEinsum = "ebe7aa44-baf0-506c-a96f-8464559b3922"
 QuantumTags = "2406e0f5-9e5e-4424-966e-8bc9c78b228f"
 Reactant = "3c362404-f566-11ee-1572-e11a4b42c853"
 cuTENSOR = "011b41b2-24ef-40a8-b3eb-fa098493e9e1"
@@ -26,6 +26,7 @@ MuscleCUDAExt = ["CUDA", "cuTENSOR"]
 MuscleChainRulesCoreExt = "ChainRulesCore"
 MuscleDaggerExt = "Dagger"
 MuscleITensorsExt = "ITensors"
+MuscleOMEinsumExt = "OMEinsum"
 MuscleQuantumTagsExt = "QuantumTags"
 MuscleReactantExt = "Reactant"
 MusclecuTensorNetExt = ["CUDA", "cuTensorNet"]
