+++ src/Operations/binary_einsum.jl
@@ -1,5 +1,4 @@
 using Adapt
-using OMEinsum: OMEinsum
 
 """
     binary_einsum(a::Tensor, b::Tensor; dims=∩(inds(a), inds(b)), out=nothing)
@@ -20,9 +19,6 @@ Perform a binary tensor contraction operation between `a` and `b` and store the 
 """
 function binary_einsum! end
 
-# TODO add a preference system for some backends
-choose_backend_rule(::typeof(binary_einsum), ::Type{<:Array}, ::Type{<:Array}) = BackendOMEinsum()
-choose_backend_rule(::typeof(binary_einsum!), ::Type{<:Array}, ::Type{<:Array}, ::Type{<:Array}) = BackendOMEinsum()
 function binary_einsum(a::Tensor, b::Tensor; dims=(∩(inds(a), inds(b))), out=nothing)
     inds_sum = ∩(dims, inds(a), inds(b))
 
@@ -58,28 +54,5 @@ function binary_einsum!(c::Tensor, a::Tensor, b::Tensor)
     end
 
     binary_einsum!(backend, data_c, inds(c), data_a, inds(a), data_b, inds(b))
-    return c
-end
-
-# backend implementations
-## `OMEinsum`
-function binary_einsum(::BackendOMEinsum, inds_c, a, inds_a, b, inds_b)
-    size_dict = Dict{Index,Int}()
-    for (ind, ind_size) in Iterators.flatten([inds_a .=> size(a), inds_b .=> size(b)])
-        size_dict[ind] = ind_size
-    end
-
-    c = OMEinsum.get_output_array((a, b), Int[size_dict[i] for i in inds_c], false)
-    OMEinsum.einsum!((inds_a, inds_b), inds_c, (a, b), c, true, false, size_dict)
-    return c
-end
-
-function binary_einsum!(::BackendOMEinsum, c, inds_c, a, inds_a, b, inds_b)
-    size_dict = Dict{Index,Int}()
-    for (ind, ind_size) in Iterators.flatten([inds_a .=> size(a), inds_b .=> size(b)])
-        size_dict[ind] = ind_size
-    end
-
-    OMEinsum.einsum!((inds_a, inds_b), inds_c, (a, b), c, true, false, size_dict)
     return c
 end
